backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
numpy
pillow
